requirements.txt
# Web应用框架
streamlit>=1.24.0
streamlit-echarts>=0.4.0

# 数据处理和分析
numpy>=1.24.3
pandas>=1.5.3
scipy>=1.10.0

# 深度学习框架
torch>=2.0.1
torchvision>=0.10.0
torchaudio>=0.9.0

# 数据可视化
plotly>=5.18.0
matplotlib>=3.7.1
seaborn>=0.12.2

# 机器学习和评估
scikit-learn>=1.2.2

# 技术分析指标
ta>=0.10.0

# 统计分析和特征选择
statsmodels>=0.13.5
pmdarima>=2.0.0

# 配置文件管理
pyyaml>=6.0

# 其他依赖
python-dateutil==2.8.2
tqdm>=4.66.0
typing-extensions==4.9.0
requests==2.31.0
yfinance>=0.2.36
lightgbm>=4.1.0
xgboost>=2.0.0
optuna>=3.4.0
joblib>=1.3.0
plotly-express>=0.4.0
jupyter>=1.0.0
pillow>=10.0.0
prophet>=1.1.1 